All endpoints correctly executed

+++ src/main/java/com/utn/udee/model/projections/UserProjection.java
@@ -11,6 +11,6 @@ public interface UserProjection {
 
     String getUsername();
 
-    Integer getSum();
+    Float getSum();
 
 }
